Restructuring of tests
