--- packages/graph-sdk/tests/src/bundle.test.js
@@ -1,14 +1,0 @@
-import '../../dist/bundle.js';
-
-test('export', async () => {
-  const { alpaca } = window;
-  expect(alpaca).not.toBeUndefined();
-  expect(alpaca.Client).not.toBeUndefined();
-
-  const apiKey = 'pk.123';
-  alpaca.apiKey = apiKey;
-
-  const client = new alpaca.Client();
-  expect(client.apiKey).toBe(apiKey);
-  expect(typeof client.query).toBe('function');
-});
